Various fix for the project creation.

+++ cookiecutters/tasks.py
@@ -5,7 +5,7 @@ import json
 from django.conf import settings
 from gittle import Gittle
 from celery import task
-# from cookiecutter.generate import generate_files
+from cookiecutter.generate import generate_files
 from baker import utils
 
 
@@ -36,13 +36,21 @@ def exec_cookiecutter(cookie, options, user_id=None, use_github=True):
         assert use_github is not True
 
     out = os.path.join(settings.COOKIECUTTERS_TMP, "user_{0}".format(user_id))
+
     try:
+        os.makedirs(out)
+
+        print out
+
         generate_files(cookie.repo_path, {'cookiecutter': options}, out)
+
         if use_github:
             repo = utils.create_repository(user_id, options['repo_name'])
+
             utils.push_directory_to_repo(os.path.join(out, options['repo_name']), repo)
         else:
             # make zippone
             pass
     finally:
-        shutil.rmtree(out)
+        if os.path.exists(out):
+            shutil.rmtree(out)
